+++ src/Symfony/Component/PasswordHasher/Hasher/PasswordHasherFactoryInterface.php
@@ -12,6 +12,7 @@
 namespace Symfony\Component\PasswordHasher\Hasher;
 
 use Symfony\Component\PasswordHasher\PasswordHasherInterface;
+use Symfony\Component\Security\Core\User\PasswordAuthenticatedUserInterface;
 use Symfony\Component\Security\Core\User\UserInterface;
 
 /**
@@ -25,7 +26,7 @@ interface PasswordHasherFactoryInterface
     /**
      * Returns the password hasher to use for the given user.
      *
-     * @param UserInterface|string $user A UserInterface instance or a class name
+     * @param PasswordAuthenticatedUserInterface|UserInterface|string $user A PasswordAuthenticatedUserInterface/UserInterface instance or a class name
      *
      * @throws \RuntimeException When no password hasher could be found for the user
      */